Begin migration to persistent projects
