+++ app/package.json
@@ -46,7 +46,8 @@
     "react-router-dom": "^6.12.1",
     "reactflow": "^11.7.0",
     "zod": "^3.21.4",
-    "zustand": "^4.3.7"
+    "zustand": "^4.3.7",
+    "db": "workspace:*"
   },
   "devDependencies": {
     "@types/cookies": "^0.7.7",
