+++ db/prisma/schema.prisma
@@ -13,7 +13,7 @@ datasource db {
 model Project {
   id         String   @id
   name       String
-  content    String
+  content    Json
   userId     String
   createdAt  DateTime @default(now())
   updatedAt  DateTime @updatedAt
